+++ js/CardList.js
@@ -1,5 +1,7 @@
 import {ElementFabric} from './ElementFabric.js';
-import {LocalStore} from './LocalStore.js'
+import {LocalStore} from './LocalStore.js';
+import {Bag} from './Bag.js';
+const bag = new Bag('#modalBag');
 const elementFabric = new ElementFabric();
 const localStore = new LocalStore();
 
@@ -19,6 +21,28 @@ export class CardList {
       }
     }));
 
+  }
+
+  _findSelectedIDs(selector) {
+    const $cardList = document.querySelector(selector);
+    const $selectedItems = [...$cardList.children]
+                              .filter(child => child.querySelector('input')?.checked);
+
+    const selectedIds = $selectedItems.map($item => +$item.id);
+    return selectedIds;
+  }
+
+  showSelectedGoods(selector) {
+    const selectedIDs = this._findSelectedIDs(selector);
+
+    localStore.saveToLocalStorage(selectedIDs);
+    let ls = localStore.getFromLocalStorage();
+
+    if (!ls || !ls.length) {
+      return
+    }
+
+    bag.renderBag(ls) && bag.showModal();
   }
 
   handlerForInput() {
