+++ js/LocalStore.js
@@ -1,6 +1,3 @@
-import {Bag} from './Bag.js';
-const bag = new Bag('#modalBag');
-
 export class LocalStore {
   constructor() {
     this.saveToLocalStorage = this.saveToLocalStorage.bind(this);
@@ -9,11 +6,9 @@ export class LocalStore {
   getFromLocalStorage() {
     const ls = JSON.parse(localStorage.getItem('selectedBeers'));
 
-    if (!ls || !ls.length) {
-      return
+    if (ls && ls.length) {
+      return ls
     }
-
-    return ls;
   }
 
   saveToLocalStorage(ids) {
@@ -34,34 +29,12 @@ export class LocalStore {
       return
     }
 
-    const index = ls.findIndex(item => item == id);
+    const index = ls.findIndex(item => item === id);
 
-    if (~index) {
+    if (index !== -1) {
       ls.splice(index, 1);
       localStorage.setItem('selectedBeers', JSON.stringify(ls));
     }
-  }
-
-  _findSelectedIDs(selector) {
-    const $cardList = document.querySelector(selector);
-    const $selectedItems = [...$cardList.children]
-                              .filter(child => child.querySelector('input')?.checked);
-
-    const selectedIds = $selectedItems.map($item => +$item.id);
-    return selectedIds;
-  }  
-
-  showSelectedGoods(selector) {
-    const selectedIDs = this._findSelectedIDs(selector);
-
-    this.saveToLocalStorage(selectedIDs);
-    let ls = this.getFromLocalStorage();
-
-    if (!ls || !ls.length) {
-      return
-    }
-
-    bag.renderBag(ls) && bag.showModal();
   }
 
   clearSelectedGoods(selector) {
